make the group movable
