+++ packages/webamp-modern-2/src/skin/makiClasses/Button.ts
@@ -115,6 +115,7 @@ export default class Button extends GuiObj {
   }
 
   _handleMouseDown(e: MouseEvent) {
+    // e.stopPropagation();
     // buttonToggle will handle it
   }
 
@@ -122,6 +123,7 @@ export default class Button extends GuiObj {
     super.draw();
     this._div.setAttribute("data-obj-name", "Button");
     this._div.classList.add("webamp--img");
+    this._div.style.pointerEvents = 'auto'
     this._renderBackground();
   }
