+++ packages/webamp-modern-2/src/skin/makiClasses/Layout.ts
@@ -157,5 +157,6 @@ export default class Layout extends Group {
   draw() {
     super.draw();
     this._div.setAttribute("data-obj-name", "Layout");
+    // this._div.style.pointerEvents = 'auto'
   }
 }
